# Optimize vfolder query on directory context