--- a/pootle/apps/virtualfolder/helpers.py
+++ b/pootle/apps/virtualfolder/helpers.py
@@ -7,6 +7,8 @@
 # or later license. See the LICENSE file for a copy of the license and the
 # AUTHORS file for copyright and authorship information.
 
+from pootle_app.models import Directory
+
 from .models import VirtualFolderTreeItem
 
 
@@ -46,3 +48,16 @@ def extract_vfolder_from_path(request_path):
         return None, request_path
     else:
         return vfti.vfolder, vfti.directory.pootle_path
+
+
+def vftis_for_child_dirs(directory):
+    """
+    Returns the vfoldertreeitems for a directory's child directories
+    """
+    child_dir_pks = [
+        child.pk
+        for child
+        in directory.children
+        if isinstance(child, Directory)]
+    return VirtualFolderTreeItem.objects.filter(
+        directory__pk__in=child_dir_pks)