--- a/pootle/core/browser.py
+++ b/pootle/core/browser.py
@@ -92,13 +92,7 @@ def make_generic_item(path_obj, **kwargs):
     }
 
 
-def make_directory_item(directory):
-    filters = {}
-
-    if directory.has_vfolders:
-        # The directory has virtual folders, so append priority sorting to URL.
-        filters['sort'] = 'priority'
-
+def make_directory_item(directory, **filters):
     item = make_generic_item(directory, **filters)
     item.update({
         'icon': 'folder',
@@ -171,19 +165,3 @@ def make_project_list_item(project):
         'title': project.fullname,
     })
     return item
-
-
-def get_children(directory):
-    """Returns a list of children directories and stores for this
-    ``directory``.
-
-    The elements of the list are dictionaries which keys are populated after
-    in the templates.
-    """
-    directories = [make_directory_item(child_dir)
-                   for child_dir in directory.child_dirs.live().iterator()]
-
-    stores = [make_store_item(child_store)
-              for child_store in directory.child_stores.live().iterator()]
-
-    return directories + stores