This finds out which subdirs have vfolders in a single db call.

My db timings for this optimization:

/af/firefox/

OLD: 22 queries run, total 0.022 seconds
NEW: 13 queries run, total 0.013 seconds

/af/firefox/browser/

OLD: 20 queries run, total 0.02 seconds
NEW: 12 queries run, total 0.012 seconds